refactor(cozy-harvest-lib): Refactor KonnectorModal
