+++ packages/cozy-harvest-lib/src/components/KonnectorModal/CreateAccountButton.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { translate, Button } from 'cozy-ui/transpiled/react/'
+
+const CreateAccount = translate()(({ createAction, t }) => {
+  return (
+    <Button
+      subtle
+      size={'small'}
+      className={'u-m-half'}
+      onClick={createAction}
+      label={t('modal.konnector.create_account')}
+    />
+  )
+})
+
+export default CreateAccount
